refactor(core): move core/utils.hpp into core/utils/
Matches the directory-aggregator convention (core/core.hpp,
application/application.hpp); resolves utils.hpp coexisting beside the
core/utils/ folder. Updates all 7 includers.
