+++ include/mori/core/core.hpp
@@ -22,7 +22,7 @@
 #pragma once
 
 #include "mori/core/transport/rdma/rdma.hpp"
-#include "mori/core/utils.hpp"
+#include "mori/core/utils/utils.hpp"
 
 #if defined(__HIPCC__) || defined(__CUDACC__)
 #include "mori/core/transport/p2p/p2p.hpp"
