+++ include/mori/core/transport/rdma/providers/mlx5/mlx5_device_primitives.hpp
@@ -29,7 +29,7 @@
 #include "mori/core/transport/rdma/device_primitives.hpp"
 #include "mori/core/transport/rdma/providers/mlx5/mlx5_defs.hpp"
 #include "mori/core/transport/rdma/utils.hpp"
-#include "mori/core/utils.hpp"
+#include "mori/core/utils/utils.hpp"
 
 namespace mori {
 namespace core {
